Extract reusable useShowStopAreaOnMap helper
Additionally fix data loading bug, just like was done for other open
on map buttons in: eb9e4d0249e66ab9008660a4f4aa3087ebd386b1
