+++ ui/src/components/stop-registry/stop-areas/stop-area-details/components/StopAreaMinimap.tsx
@@ -1,49 +1,12 @@
 import { FC } from 'react';
-// Don't forget to remove the image from the repo!
 import { useTranslation } from 'react-i18next';
 import { twMerge } from 'tailwind-merge';
-import { StopAreaDetailsFragment } from '../../../../../generated/graphql';
-import {
-  useAppDispatch,
-  useFilterStops,
-  useMapQueryParams,
-  useObservationDateQueryParam,
-} from '../../../../../hooks';
-import {
-  FilterType,
-  resetMapState,
-  setSelectedMapStopAreaIdAction,
-} from '../../../../../redux';
 import { mapLngLatToPoint } from '../../../../../utils';
 import { SlimSimpleButton } from '../../../stops/stop-details/layout';
+import { useShowStopAreaOnMap } from '../../../utils';
+// Don't forget to remove the image from the repo!
 import placeholderBg from '../PlaceholderMapFragment.png';
 import { StopAreaComponentProps } from './StopAreaComponentProps';
-
-function useShowOnMap() {
-  const dispatch = useAppDispatch();
-  const { observationDate } = useObservationDateQueryParam();
-  const { openMapWithParameters } = useMapQueryParams();
-  const { toggleFunction } = useFilterStops();
-  const toggleShowAllStops = toggleFunction(FilterType.ShowAllBusStops);
-
-  return (area: StopAreaDetailsFragment) => {
-    dispatch(resetMapState()).then(() => {
-      dispatch(setSelectedMapStopAreaIdAction(area.id ?? undefined));
-      toggleShowAllStops(false);
-
-      const point = mapLngLatToPoint(area.geometry?.coordinates ?? []);
-      openMapWithParameters({
-        viewPortParams: {
-          latitude: point.latitude,
-          longitude: point.longitude,
-          zoom: 18,
-        },
-        observationDate,
-        displayedRouteParams: {},
-      });
-    });
-  };
-}
 
 const testIds = {
   openMapButton: 'StopAreaMinimap::openMapButton',
@@ -55,7 +18,7 @@ export const StopAreaMinimap: FC<StopAreaComponentProps> = ({
   className = '',
 }) => {
   const { t } = useTranslation();
-  const showOnMap = useShowOnMap();
+  const showOnMap = useShowStopAreaOnMap();
 
   const point = mapLngLatToPoint(area.geometry?.coordinates ?? []);
 
@@ -84,7 +47,7 @@ export const StopAreaMinimap: FC<StopAreaComponentProps> = ({
       <SlimSimpleButton
         className="absolute right-2 top-2"
         inverted
-        onClick={() => showOnMap(area)}
+        onClick={() => showOnMap(area.id ?? undefined, point)}
         testId={testIds.openMapButton}
       >
         {t('stopAreaDetails.minimap.showOnMap')}
